Setup FastAPI backend with uv and MongoDB
