+++ .env.example
@@ -1,0 +1,2 @@
+# Example environment variables
+MONGO_URI=mongodb+srv://<username>:<password>@cluster0.example.mongodb.net/mydb
