+++ backend/tests/test_placeholder.py
@@ -1,0 +1,5 @@
+import pytest
+
+
+def test_placeholder():
+    assert True
